Added helper methods for parsing function/stored procedures
The class sqlchecker.io.IOUtil does now contain helper functions for
parsing the header from a stored procedure or a function. 
These functions are also able to parse the variables and the name from a
stored procedure/function. After parsing, the information about the
function/stored procedure is returned as a token list in which the first
token contains the name of the method.

Implementing and testing this took longer than expected.