Implemented script execution and improved code style
* Implemented the MySQLQueryExecuter class. This class will connect to a
database and run execute() for every query in the query list which gets
passed to the runSQL() method of this class.
* ScriptReader class will now read the given script, connect to mysql
and execute all the queries which were read by using the
MySQLQueryExecuter class

* Implemented MySQLWrapper class which provides all the
init/close/rollback/store-result functions. This class gets extended by
every class which uses a mysql connection