Improved AbstractFileReader method visibility
The visibility of beforeReading() onReadLine() and afterReading() was
changed from public to protected. This will make those methods
inaccessible from outside of instances of classes which extend
AbstractFileReader. These methods should never be called from outside an
instance of one of those classes so "public" visibility was a bad
choice.