Fixed tag-detection (ticket #20)
If there were non-ASCII characters near a tag, then a tag might not get
detected properly. This was fixed by stripping all non-ASCII characters
(from the test string) before performing any actual checks. This is
another step of the string normalization process before the
tag-detection is done

Also added the utf8test class which tests this bug and shows that the
bug was fixed. If the bug would not be fixed, then the SQL statement
associated with the first tag would be empty ("")