Fixed 2 bugs related to the submission execution
* [SubmissionExecuter():Constructor] Fixed a bug which would cause file
path strings to have a surplus "/" in it. E.g.
data/assignment1//solution.txt instead of data/assignment1/solution.txt
While this bug did not cause any visible problems, it is still a bug
which might cause unwanted side effects
* [SolutionReader:onReadLine(line)] Fixed a bug related to the detection
of the tags= annotation in the solution file. Sometimes, these tags do
not get detected properly causing the tag list to be empty. This was
caused by non-visible non-ascii characters. As a fix, there is a
pre-processing step which would strip non-ascii characters from the line
which should be checked. For all other operations, the "original" line
variable is used